Create README.md
This all started with an issue I have getting my code to pass a VeraCode scan. This project purports to have some examples that I might use to help with this issue.

The first thing I noticed was that the code hasn't been kept up to date. The references don't exist. It did not compile etc... ARG!

The first commit just updated the code to .Net 4.5 and enabled some Nuget package restores that will let the thing compile. Im sure I will hate to have to use this library but X'ing my fingers.